fix: dynamically allocate Expr evaluation results
Previously this was all statically allocated, making for a lot of
unnecessary copies on top of dangling pointers to on-stack Values.

This became apparent when writing the ClassInstance code, since this was
the first time a variable relied on _another_ variable (e.g
ClassInstance needs a reference to the Class it is instantiated from).

This has been fixed by dynamiccally allocating variables and their usage
within the Evaluator. This is quite an extensive change.